Move algorithm and meta sublibraries outside of core
