+++ README.md
@@ -1,6 +1,6 @@
 [![License](https://img.shields.io/badge/License-LGPL%203.0-blue.svg)](https://opensource.org/licenses/LGPL-3.0)
 ![CMake](https://github.com/fair-acc/graph-prototype/workflows/CMake/badge.svg)
-# Graph Prototype
+# GNURadio 4.0 prototype
 A small proof-of-concept for evaluating efficient [directed graph](https://en.wikipedia.org/wiki/Directed_graph)-based
 algorithms, notably required node structures, scheduling interfaces, and partial compile-time merging of
 [directed acyclic](https://en.wikipedia.org/wiki/Directed_acyclic_graph) as well as
