+++ algorithm/README.md
@@ -1,0 +1,6 @@
+# gnuradio-algorithm
+
+A math support library with math primitives to be used by different block types.
+
+Included algorithms:
+ - Fast Fourier Transform
