+++ blocks/basic/test/qa_selector.cpp
@@ -3,9 +3,10 @@
 
 #include <boost/ut.hpp>
 
+#include <gnuradio-4.0/meta/utils.hpp>
+
 #include <gnuradio-4.0/graph.hpp>
 #include <gnuradio-4.0/node.hpp>
-#include <gnuradio-4.0/utils.hpp>
 
 #include <gnuradio-4.0/basic/selector.hpp>
