+++ blocks/filter/test/CMakeLists.txt
@@ -1,1 +1,2 @@
 add_ut_test(qa_filter)
+target_link_libraries(qa_filter PRIVATE gr-filter)
