+++ blocks/fourier/test/CMakeLists.txt
@@ -1,1 +1,2 @@
 add_ut_test(qa_fourier)
+target_link_libraries(qa_fourier PRIVATE gr-fourier)
