+++ meta/README.md
@@ -1,0 +1,3 @@
+# gnuradio-meta
+
+A small metaprogramming and utility library for use by different parts of GNURadio.
