+++ meta/test/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_ut_test(qa_traits)
+target_link_libraries(qa_traits PRIVATE gnuradio-meta)
